+++ ChangeLog.md
@@ -4,6 +4,7 @@
 
 - ldap: respect `create_users` configuration
 - Fix bug creating new priority (@balsdorf)
+- add MailTransport class to encapsulate smtp transport
 
 [3.1.10]: https://github.com/eventum/eventum/compare/v3.1.9...master